Split parameters.py into separate files
* Add a subpackage odxtools/parameters/ which contains the class definitions
  which previously were part of `parameters.py`
